the view of affinity modules optimization
