+++ build.gradle
@@ -11,6 +11,7 @@ buildscript {
     ext.butterknife_gradle_plugin_version = '9.0.0'
 
     repositories {
+//        maven { url "https://jitpack.io" }
         mavenCentral()
         mavenLocal()
         google()
