+++ research/view/src/main/java/com/dvsnier/test/view/affinity/BActivity.java
@@ -1,0 +1,45 @@
+package com.dvsnier.test.view.affinity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
+
+import com.dvsnier.test.view.R;
+import com.dvsnier.test.view.R2;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+import butterknife.OnClick;
+
+public class BActivity extends AppCompatActivity {
+
+    @BindView(R2.id.btn_affinity_a)
+    Button btnAffinityA;
+    @BindView(R2.id.btn_c)
+    Button btnC;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_b);
+        ButterKnife.bind(this);
+    }
+
+    @OnClick({R2.id.btn_affinity_a, R2.id.btn_c})
+    public void onClick(View view) {
+        Intent intent = null;
+        int viewId = view.getId();
+        if (viewId == R.id.btn_affinity_a) {
+            intent = new Intent(this, AActivity.class);
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            startActivity(intent);
+        } else if (viewId == R.id.btn_c) {
+            intent = new Intent(this, CActivity.class);
+            startActivity(intent);
+        } else {
+            // nothing to do
+        }
+    }
+}
