+++ research/view/src/main/res/layout/activity_c.xml
@@ -3,18 +3,19 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.dvsnier.testAffinity.BActivity">
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    tools:context="com.dvsnier.test.view.affinity.CActivity">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_centerInParent="true"
         android:gravity="center"
-        android:text="B Activity" />
+        android:text="C Activity"
+        tools:ignore="HardcodedText" />
 
     <Button
         android:id="@+id/btn_affinity_a"
@@ -23,13 +24,15 @@
         android:layout_alignParentTop="true"
         android:layout_centerVertical="true"
         android:layout_marginTop="25dp"
-        android:text="测试 Affinity A" />
+        android:text="测试 Affinity A"
+        tools:ignore="HardcodedText" />
 
     <Button
-        android:id="@+id/btn_c"
+        android:id="@+id/btn_affinity_b"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@+id/btn_affinity_a"
         android:layout_marginTop="25dp"
-        android:text="测试 C" />
+        android:text="测试 Affinity B"
+        tools:ignore="HardcodedText" />
 </RelativeLayout>
